Refactor: #80 CircuitBreaker를 Client로 따로 만들기
