+++ src/main/java/com/back/global/igdb/IgdbClient.java
@@ -6,7 +6,6 @@ import com.back.global.igdb.dto.IgdbGameDetailDto;
 import com.back.global.igdb.dto.IgdbGameSummaryDto;
 import com.back.global.igdb.dto.IgdbGenreDto;
 import com.back.global.igdb.exception.IgdbApiException;
-import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -46,10 +45,7 @@ public class IgdbClient {
         }
     }
 
-    @CircuitBreaker(name = "igdb", fallbackMethod = "searchGamesFallback")
     public List<IgdbGameSummaryDto> searchGames(String keyword, int limit) {
-        // IGDB Query Language (APICALYPSE)
-        // search "elden ring"; fields id,name,summary,first_release_date,cover.url; limit 10;
         String body = """
                 search "%s";
                 fields id,name,summary,first_release_date;
@@ -60,7 +56,6 @@ public class IgdbClient {
         return res == null ? List.of() : List.of(res);
     }
 
-    @CircuitBreaker(name = "igdb", fallbackMethod = "getGameDetailFallback")
     public IgdbGameDetailDto getGameDetail(long igdbId) {
         String body = """
             fields
@@ -79,7 +74,6 @@ public class IgdbClient {
         return firstOrNull(res);
     }
 
-    @CircuitBreaker(name = "igdb", fallbackMethod = "getGameNameFallback")
     public IgdbGameNameDto getGameName(long igdbId) {
         String body = """
             fields
@@ -92,7 +86,6 @@ public class IgdbClient {
         return firstOrNull(res);
     }
 
-    @CircuitBreaker(name = "igdb", fallbackMethod = "getVideoIdFallback")
     public IgdbVideoDto getVideoId(long igdbGameId) {
         String body = """
             fields
@@ -106,7 +99,6 @@ public class IgdbClient {
         return firstOrNull(res);
     }
 
-    @CircuitBreaker(name = "igdb", fallbackMethod = "getSimilarGameIdsFallback")
     public List<Long> getSimilarGameIds(long igdbId) {
         String body = """
             fields
@@ -119,7 +111,6 @@ public class IgdbClient {
         return res[0].similar_games();
     }
 
-    @CircuitBreaker(name = "igdb", fallbackMethod = "getSimilarGameBriefByIdFallback")
     public List<SimilarGameResponse> getSimilarGameBriefById(List<Long> ids) {
         List<Long> picked = ids.stream().limit(30).toList();
         if (picked.isEmpty()) return List.of();
@@ -143,7 +134,6 @@ public class IgdbClient {
         return out;
     }
 
-    @CircuitBreaker(name = "igdb", fallbackMethod = "getPopularGameIdsFallback")
     public List<IgdbPopularityPrimitiveDto> getPopularGameIds(int limit) {
         String body = """
             fields game_id,value,popularity_type;
@@ -157,7 +147,6 @@ public class IgdbClient {
         return res == null ? List.of() : List.of(res);
     }
 
-    @CircuitBreaker(name = "igdb", fallbackMethod = "getGamesByIdsFallback")
     public List<IgdbPopularGameDto> getGamesByIds(List<Long> gameIds) {
         if (gameIds == null || gameIds.isEmpty()) return List.of();
 
@@ -172,7 +161,6 @@ public class IgdbClient {
         return res == null ? List.of() : List.of(res);
     }
 
-    @CircuitBreaker(name = "igdb", fallbackMethod = "fetchGamesByIdsFallback")
     public List<GameRow> fetchGamesByIds(List<Long> idsInOrder) {
         String idList = idsInOrder.stream().map(String::valueOf).collect(Collectors.joining(","));
 
@@ -187,9 +175,8 @@ public class IgdbClient {
         return games == null ? List.of() : List.of(games);
     }
 
-    @CircuitBreaker(name = "igdb", fallbackMethod = "getGameRatingFallback")
     public IgdbPopularGameDto getGameRating(long igdbId) {
-        String body = """                                                                                        
+        String body = """
           fields id, name, cover.image_id, total_rating, total_rating_count;
           where id = %d;
           limit 1;
@@ -233,64 +220,6 @@ public class IgdbClient {
         return s.replace("\"", "\\\"");
     }
 
-    // ── Circuit Breaker Fallback Methods ──
-
-    private List<IgdbGameSummaryDto> searchGamesFallback(String keyword, int limit, Throwable t) {
-        log.warn("IGDB Circuit Breaker OPEN – searchGames fallback, cause: {}", t.getMessage());
-        return List.of();
-    }
-
-    private IgdbGameDetailDto getGameDetailFallback(long igdbId, Throwable t) {
-        log.warn("IGDB Circuit Breaker OPEN – getGameDetail fallback, cause: {}", t.getMessage());
-        return null;
-    }
-
-    private IgdbGameNameDto getGameNameFallback(long igdbId, Throwable t) {
-        log.warn("IGDB Circuit Breaker OPEN – getGameName fallback, cause: {}", t.getMessage());
-        return null;
-    }
-
-    private IgdbVideoDto getVideoIdFallback(long igdbGameId, Throwable t) {
-        log.warn("IGDB Circuit Breaker OPEN – getVideoId fallback, cause: {}", t.getMessage());
-        return null;
-    }
-
-    private List<Long> getSimilarGameIdsFallback(long igdbId, Throwable t) {
-        log.warn("IGDB Circuit Breaker OPEN – getSimilarGameIds fallback, cause: {}", t.getMessage());
-        return List.of();
-    }
-
-    private List<SimilarGameResponse> getSimilarGameBriefByIdFallback(List<Long> ids, Throwable t) {
-        log.warn("IGDB Circuit Breaker OPEN – getSimilarGameBriefById fallback, cause: {}", t.getMessage());
-        return List.of();
-    }
-
-    private List<IgdbPopularityPrimitiveDto> getPopularGameIdsFallback(int limit, Throwable t) {
-        log.warn("IGDB Circuit Breaker OPEN – getPopularGameIds fallback, cause: {}", t.getMessage());
-        return List.of();
-    }
-
-    private List<IgdbPopularGameDto> getGamesByIdsFallback(List<Long> gameIds, Throwable t) {
-        log.warn("IGDB Circuit Breaker OPEN – getGamesByIds fallback, cause: {}", t.getMessage());
-        return List.of();
-    }
-
-    private List<GameRow> fetchGamesByIdsFallback(List<Long> idsInOrder, Throwable t) {
-        log.warn("IGDB Circuit Breaker OPEN – fetchGamesByIds fallback, cause: {}", t.getMessage());
-        return List.of();
-    }
-
-    private IgdbPopularGameDto getGameRatingFallback(long igdbId, Throwable t) {
-        log.warn("IGDB Circuit Breaker OPEN – getGameRating fallback, cause: {}", t.getMessage());
-        return null;
-    }
-
-    private List<IgdbGenreDto> fetchGenresFallback(Throwable t) {
-        log.warn("IGDB Circuit Breaker OPEN – fetchGenres fallback, cause: {}", t.getMessage());
-        return List.of();
-    }
-
-    @CircuitBreaker(name = "igdb", fallbackMethod = "fetchGenresFallback")
     public List<IgdbGenreDto> fetchGenres() {
         String body = """
         fields id,name;
